Aggiunti autowiring di jdbc tempale e simple messaging. Esempio di controller di api
